requirements.txt
asgiref==3.2.9
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
django-import-export==2.2.0
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
idna==2.10
instaloader==4.5.5
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.1.5
pylint==2.5.3
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
urllib3==1.26.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
langdetect=1.0.9
